+++ TPI/juegoDeLaVida/CMakeLists.txt
@@ -9,7 +9,7 @@ include(FindSDL2)
 add_subdirectory(lib/googletest-master)
 include_directories(lib/googletest-master/googletest/include)
 
-set(SOURCE_FILES_TEST main.cpp solucion.h solucion.cpp definiciones.h tests/soloBloquesTEST.cpp tests/esValidoTEST.cpp tests/posicionesVivasTEST.cpp tests/densidadPoblacionTEST.cpp tests/evolucionDePosicionTEST.cpp tests/evolucionMultipleTEST.cpp tests/evolucionToroideTEST.cpp tests/esPeriodicoTEST.cpp tests/primosLejanosTEST.cpp tests/seleccionNaturalTEST.cpp tests/fusionarTEST.cpp tests/vistaTrasladadaTEST.cpp tests/enCrecimientoTEST.cpp toroidUtils.cpp toroidUtils.h vectorUtils.cpp vectorUtils.h)
+set(SOURCE_FILES_TEST main.cpp solucion.h solucion.cpp definiciones.h tests/soloBloquesTEST.cpp tests/esValidoTEST.cpp tests/posicionesVivasTEST.cpp tests/densidadPoblacionTEST.cpp tests/evolucionDePosicionTEST.cpp tests/evolucionMultipleTEST.cpp tests/evolucionToroideTEST.cpp tests/esPeriodicoTEST.cpp tests/primosLejanosTEST.cpp tests/seleccionNaturalTEST.cpp tests/fusionarTEST.cpp tests/vistaTrasladadaTEST.cpp tests/enCrecimientoTEST.cpp vectorUtils.h toroidUtils.cpp toroidUtils.h)
 add_executable(juegoDeLaVidaTest ${SOURCE_FILES_TEST})
 
 set_target_properties(juegoDeLaVidaTest PROPERTIES COMPILE_FLAGS "-g --coverage")
@@ -18,7 +18,7 @@ set_target_properties(juegoDeLaVidaTest PROPERTIES LINK_FLAGS "-g --coverage")
 target_link_libraries(juegoDeLaVidaTest gtest gtest_main)
 
 if(SDL2_FOUND)
-	set(SOURCE_FILES_GUI juegoDeLaVida.cpp solucion.h solucion.cpp definiciones.h es.h es.cpp maxiopt.h maxiopt.cpp toroidUtils.cpp toroidUtils.h vectorUtils.cpp vectorUtils.h)
+	set(SOURCE_FILES_GUI juegoDeLaVida.cpp solucion.h solucion.cpp definiciones.h es.h es.cpp maxiopt.h maxiopt.cpp vectorUtils.h toroidUtils.cpp toroidUtils.h)
 	add_executable(juegoDeLaVida ${SOURCE_FILES_GUI})
 
 	set_target_properties(juegoDeLaVida PROPERTIES COMPILE_FLAGS "-O3")
